+++ oshi-core/src/main/java/oshi/hardware/HWDiskStore.java
@@ -30,7 +30,7 @@ import java.io.Serializable;
  *
  * @author enrico[dot]bianchi[at]gmail[dot]com
  */
-public class HWDiskStore implements Serializable {
+public class HWDiskStore implements Serializable, Comparable<HWDiskStore> {
 
     private static final long serialVersionUID = 1L;
 
@@ -43,12 +43,13 @@ public class HWDiskStore implements Serializable {
     private long writes;
     private long writeBytes;
     private long transferTime;
+    private HWPartition[] partitions;
 
     /**
      * Create an object with empty/default values
      */
     public HWDiskStore() {
-        this("", "", "", 0L, 0L, 0L, 0L, 0L, 0L);
+        this("", "", "", 0L, 0L, 0L, 0L, 0L, 0L, new HWPartition[0]);
     }
 
     /**
@@ -72,9 +73,11 @@ public class HWDiskStore implements Serializable {
      *            Number of bytes written to the disk
      * @param transferTime
      *            milliseconds spent reading or writing to the disk
+     * @param partitions
+     *            Partitions on this disk
      */
     public HWDiskStore(String name, String model, String serial, long size, long reads, long readBytes, long writes,
-            long writeBytes, long transferTime) {
+            long writeBytes, long transferTime, HWPartition[] partitions) {
         setName(name);
         setModel(model);
         setSerial(serial);
@@ -84,6 +87,7 @@ public class HWDiskStore implements Serializable {
         setWrites(writes);
         setWriteBytes(writeBytes);
         setTransferTime(transferTime);
+        setPartitions(partitions);
     }
 
     /**
@@ -147,6 +151,13 @@ public class HWDiskStore implements Serializable {
      */
     public long getTransferTime() {
         return transferTime;
+    }
+
+    /**
+     * @return Returns the partitions on this drive.
+     */
+    public HWPartition[] getPartitions() {
+        return partitions;
     }
 
     /**
@@ -219,5 +230,22 @@ public class HWDiskStore implements Serializable {
      */
     public void setTransferTime(long transferTime) {
         this.transferTime = transferTime;
+    }
+
+    /**
+     * @param partitions
+     *            The partitions to set.
+     */
+    public void setPartitions(HWPartition[] partitions) {
+        this.partitions = partitions;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public int compareTo(HWDiskStore store) {
+        // Naturally sort by device name
+        return this.getName().compareTo(store.getName());
     }
 }