+++ oshi-core/src/main/java/oshi/hardware/platform/mac/MacDisks.java
@@ -19,7 +19,12 @@
 package oshi.hardware.platform.mac;
 
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,10 +34,13 @@ import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.PointerByReference;
 
 import oshi.hardware.HWDiskStore;
+import oshi.hardware.HWPartition;
 import oshi.hardware.common.AbstractDisks;
 import oshi.jna.platform.mac.CoreFoundation;
+import oshi.jna.platform.mac.CoreFoundation.CFBooleanRef;
 import oshi.jna.platform.mac.CoreFoundation.CFDictionaryRef;
 import oshi.jna.platform.mac.CoreFoundation.CFMutableDictionaryRef;
+import oshi.jna.platform.mac.CoreFoundation.CFNumberRef;
 import oshi.jna.platform.mac.CoreFoundation.CFStringRef;
 import oshi.jna.platform.mac.DiskArbitration;
 import oshi.jna.platform.mac.DiskArbitration.DADiskRef;
@@ -54,8 +62,11 @@ public class MacDisks extends AbstractDisks {
 
     private static final Logger LOG = LoggerFactory.getLogger(MacDisks.class);
 
+    private static final Map<String, String> mountPointMap = new HashMap<>();
+
     @Override
     public HWDiskStore[] getDisks() {
+        mountPointMap.clear();
         List<HWDiskStore> result = new ArrayList<>();
 
         // Use statfs to find all drives
@@ -65,26 +76,34 @@ public class MacDisks extends AbstractDisks {
         // Fill array with results
         SystemB.INSTANCE.getfsstat64(fs, numfs * new Statfs().size(), SystemB.MNT_NOWAIT);
 
-        // Open a DiskArbitration session to get model and size of disks
+        // Open a DiskArbitration session
         DASessionRef session = DiskArbitration.INSTANCE.DASessionCreate(CfUtil.ALLOCATOR);
         if (session == null) {
             LOG.error("Unable to open session to DiskArbitration framework.");
             return new HWDiskStore[0];
         }
 
+        // Create a set in case we have multiple partitions mounted
+        Set<String> bsdNames = new HashSet<String>();
         // Iterate all mounted file systems
         for (Statfs f : fs) {
+            String mntFrom = new String(f.f_mntfromname).trim();
+            // OS X registry uses the BSD Name, e.g. disk0, disk1, etc.
+            // Strip off the partition # to get base disk name
+            String[] split = mntFrom.split("/dev/|s\\d+");
+            if (split.length > 1) {
+                bsdNames.add(split[1]);
+            }
+            mountPointMap.put(mntFrom.replace("/dev/", ""), new String(f.f_mntonname).trim());
+        }
+        // Now iterate the bsdNames
+        for (String bsdName : bsdNames) {
             String model = "";
             String serial = "";
             long size = 0L;
             long xferTime = 0L;
-            // Get a reference to the disk - only matching /dev/disk*s2
-            String[] split = new String(f.f_mntfromname).trim().split("/dev/|s2");
-            if (split.length < 2) {
-                continue;
-            }
-            // OS X registry uses the BSD Name, e.g. disk0, disk1, etc.
-            String bsdName = split[1];
+
+            // Get a reference to the disk - only matching /dev/disk*
             String path = "/dev/" + bsdName;
 
             // Get the DiskArbitration dictionary for this disk, which has model
@@ -139,11 +158,12 @@ public class MacDisks extends AbstractDisks {
                 }
                 CfUtil.release(disk);
 
-                //
+                // If empty, ignore
                 if (size <= 0) {
                     continue;
                 }
-                HWDiskStore diskStore = new HWDiskStore(bsdName, model.trim(), serial.trim(), size, 0L, 0L, 0L, 0L, 0L);
+                HWDiskStore diskStore = new HWDiskStore(bsdName, model.trim(), serial.trim(), size, 0L, 0L, 0L, 0L, 0L,
+                        new HWPartition[0]);
 
                 // Now look up the device using the BSD Name to get its
                 // statistics
@@ -204,6 +224,84 @@ public class MacDisks extends AbstractDisks {
                             } else {
                                 LOG.error("Unable to find block storage driver properties for {}", bsdName);
                             }
+                            // Now get partitions for this disk.
+                            List<HWPartition> partitions = new ArrayList<>();
+                            if (IOKit.INSTANCE.IORegistryEntryCreateCFProperties(drive, propsPtr, CfUtil.ALLOCATOR,
+                                    0) == 0) {
+                                CFMutableDictionaryRef properties = new CFMutableDictionaryRef();
+                                properties.setPointer(propsPtr.getValue());
+                                // Partitions will match BSD Unit property
+                                Pointer p = CoreFoundation.INSTANCE.CFDictionaryGetValue(properties,
+                                        CfUtil.getCFString("BSD Unit"));
+                                CFNumberRef bsdUnit = new CFNumberRef();
+                                bsdUnit.setPointer(p);
+                                // Whole disk has 'true' for Whole and 'false'
+                                // for leaf; store the boolean true
+                                p = CoreFoundation.INSTANCE.CFDictionaryGetValue(properties,
+                                        CfUtil.getCFString("Whole"));
+                                CFBooleanRef leaf = new CFBooleanRef();
+                                leaf.setPointer(p);
+                                // create a matching dict for BSD Unit
+                                CFMutableDictionaryRef propertyDict = CoreFoundation.INSTANCE
+                                        .CFDictionaryCreateMutable(CfUtil.ALLOCATOR, 0, null, null);
+                                CoreFoundation.INSTANCE.CFDictionarySetValue(propertyDict,
+                                        CfUtil.getCFString("BSD Unit"), bsdUnit);
+                                CoreFoundation.INSTANCE.CFDictionarySetValue(propertyDict, CfUtil.getCFString("Leaf"),
+                                        leaf);
+                                matchingDict = CoreFoundation.INSTANCE.CFDictionaryCreateMutable(CfUtil.ALLOCATOR, 0,
+                                        null, null);
+                                CoreFoundation.INSTANCE.CFDictionarySetValue(matchingDict,
+                                        CfUtil.getCFString("IOPropertyMatch"), propertyDict);
+
+                                // search for all IOservices that match the
+                                // BSD Unit with leaf=true; these are partitions
+                                IntByReference serviceIterator = new IntByReference();
+                                IOKitUtil.getMatchingServices(matchingDict, serviceIterator);
+                                // getMatchingServices releases matchingDict
+                                CfUtil.release(properties);
+                                CfUtil.release(propertyDict);
+                                int sdService = IOKit.INSTANCE.IOIteratorNext(serviceIterator.getValue());
+                                while (sdService != 0) {
+                                    // look up the BSD Name
+                                    String partBsdName = IOKitUtil.getIORegistryStringProperty(sdService, "BSD Name");
+                                    String name = partBsdName;
+                                    String type = "";
+                                    // Get the DiskArbitration dictionary for
+                                    // this partition
+                                    disk = DiskArbitration.INSTANCE.DADiskCreateFromBSDName(CfUtil.ALLOCATOR, session,
+                                            partBsdName);
+                                    if (disk != null) {
+                                        diskInfo = DiskArbitration.INSTANCE.DADiskCopyDescription(disk);
+                                        if (diskInfo != null) {
+                                            // get volume name from its key
+                                            Pointer volumePtr = CoreFoundation.INSTANCE.CFDictionaryGetValue(diskInfo,
+                                                    CfUtil.getCFString("DAMediaName"));
+                                            type = CfUtil.cfPointerToString(volumePtr);
+                                            volumePtr = CoreFoundation.INSTANCE.CFDictionaryGetValue(diskInfo,
+                                                    CfUtil.getCFString("DAVolumeName"));
+                                            name = CfUtil.cfPointerToString(volumePtr);
+
+                                            CfUtil.release(diskInfo);
+                                        }
+                                        CfUtil.release(disk);
+                                    }
+                                    partitions.add(new HWPartition(partBsdName, name, type,
+                                            IOKitUtil.getIORegistryStringProperty(sdService, "UUID"),
+                                            IOKitUtil.getIORegistryLongProperty(sdService, "Size"),
+                                            IOKitUtil.getIORegistryIntProperty(sdService, "BSD Major"),
+                                            IOKitUtil.getIORegistryIntProperty(sdService, "BSD Minor"),
+                                            mountPointMap.getOrDefault(partBsdName, "")));
+                                    IOKit.INSTANCE.IOObjectRelease(sdService);
+                                    // iterate
+                                    sdService = IOKit.INSTANCE.IOIteratorNext(serviceIterator.getValue());
+                                }
+                                IOKit.INSTANCE.IOObjectRelease(serviceIterator.getValue());
+
+                            } else {
+                                LOG.error("Unable to find properties for {}", bsdName);
+                            }
+                            Collections.sort(partitions);
+                            diskStore.setPartitions(partitions.toArray(new HWPartition[partitions.size()]));
                             IOKit.INSTANCE.IOObjectRelease(parent.getValue());
                         } else {
                             LOG.error("Unable to find IOMedia device or parent for ", bsdName);
@@ -217,6 +315,7 @@ public class MacDisks extends AbstractDisks {
         }
         // Close DA session
         CfUtil.release(session);
+        Collections.sort(result);
         return result.toArray(new HWDiskStore[result.size()]);
     }