+++ oshi-core/src/main/java/oshi/hardware/platform/unix/solaris/SolarisUsbDevice.java
@@ -27,6 +27,7 @@ import java.util.Map;
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractUsbDevice;
 import oshi.util.ExecutingCommand;
+import oshi.util.ParseUtil;
 
 public class SolarisUsbDevice extends AbstractUsbDevice {
 
@@ -122,11 +123,11 @@ public class SolarisUsbDevice extends AbstractUsbDevice {
             // approrpriate variables to maps.
             line = line.trim();
             if (line.startsWith("model:")) {
-                nameMap.put(key, getPrtconfString(line));
+                nameMap.put(key, ParseUtil.getSingleQuoteStringValue(line));
             } else if (line.startsWith("name:")) {
                 // Name is backup for model if model doesn't exist, so only
                 // put if key doesn't yet exist
-                nameMap.putIfAbsent(key, getPrtconfString(line));
+                nameMap.putIfAbsent(key, ParseUtil.getSingleQuoteStringValue(line));
             } else if (line.startsWith("vendor-id:")) {
                 // Format: vendor-id: 00008086
                 if (line.length() > 4) {
@@ -140,7 +141,7 @@ public class SolarisUsbDevice extends AbstractUsbDevice {
             } else if (line.startsWith("device_type:")) {
                 // Name is backup for model if model doesn't exist, so only
                 // put if key doesn't yet exist
-                deviceTypeMap.putIfAbsent(key, getPrtconfString(line));
+                deviceTypeMap.putIfAbsent(key, ParseUtil.getSingleQuoteStringValue(line));
             }
         }
 
@@ -153,21 +154,6 @@ public class SolarisUsbDevice extends AbstractUsbDevice {
             }
         }
         return controllerDevices.toArray(new UsbDevice[controllerDevices.size()]);
-    }
-
-    /**
-     * Parses a string key = 'value' (string)
-     * 
-     * @param line
-     *            The entire string from lshal
-     * @return the value contained between single tick marks
-     */
-    private static String getPrtconfString(String line) {
-        String[] split = line.split("'");
-        if (split.length < 2) {
-            return "";
-        }
-        return split[1];
     }
 
     /**