+++ oshi-json/src/main/java/oshi/json/hardware/impl/HardwareAbstractionLayerImpl.java
@@ -30,6 +30,7 @@ import oshi.json.hardware.CentralProcessor;
 import oshi.json.hardware.Display;
 import oshi.json.hardware.GlobalMemory;
 import oshi.json.hardware.HWDiskStore;
+import oshi.json.hardware.HWPartition;
 import oshi.json.hardware.HardwareAbstractionLayer;
 import oshi.json.hardware.NetworkIF;
 import oshi.json.hardware.PowerSource;
@@ -109,9 +110,17 @@ public class HardwareAbstractionLayerImpl extends AbstractOshiJsonObject impleme
         oshi.hardware.HWDiskStore[] ds = this.hal.getDiskStores();
         HWDiskStore[] diskStores = new HWDiskStore[ds.length];
         for (int i = 0; i < ds.length; i++) {
+            HWPartition[] partitions = new HWPartition[ds[i].getPartitions().length];
+            for (int j = 0; j < partitions.length; j++) {
+                partitions[j] = new HWPartition(ds[i].getPartitions()[j].getIdentification(),
+                        ds[i].getPartitions()[j].getName(), ds[i].getPartitions()[j].getType(),
+                        ds[i].getPartitions()[j].getUuid(), ds[i].getPartitions()[j].getSize(),
+                        ds[i].getPartitions()[j].getMajor(), ds[i].getPartitions()[j].getMinor(),
+                        ds[i].getPartitions()[j].getMountPoint());
+            }
             diskStores[i] = new HWDiskStore(ds[i].getName(), ds[i].getModel(), ds[i].getSerial(), ds[i].getSize(),
                     ds[i].getReads(), ds[i].getReadBytes(), ds[i].getWrites(), ds[i].getWriteBytes(),
-                    ds[i].getTransferTime());
+                    ds[i].getTransferTime(), partitions);
         }
         return diskStores;
     }